Convert UserMentionsDropdown styles to SCSS module
