--- packages/slate-editor/src/modules/editor-v4-user-mentions/components/UserMentionsDropdown/index.ts
@@ -1,1 +1,0 @@
-export { UserMentionsDropdown } from './UserMentionsDropdown';
