--- packages/slate-editor/src/modules/editor-v4-user-mentions/components/UserMentionsDropdownOption/index.ts
@@ -1,1 +1,0 @@
-export { UserMentionsDropdownOption } from './UserMentionsDropdownOption';
